Use integers for actions and add reduce function
Previously the plan was to fill the action table with objects of type
Action. The problem with this plan was that the action table can grow
extremely large. To keep the table small all actions are now encoded in
integers.

The other change is the addition of the reduce function.

With these two changes, the parser is now "functional"! 🚀