Add skip action to table
In parsers it is common to have the lexer swallow tokens such as white
space or newlines, while this should actually be the responsibility of
parser. A naive way to handle this in the parser, is to keep a set of
token types that must be skipped and check every incoming token. Much
prettier is to make this part of the main loop. We have done this by
placing action type 'skip' in the entire column belonging to the token
type (terminal). Now, when the parser reads such a token, regardless of
the current state, it finds a skip action in the table and thus
continues to the next token.