Build description instead of definition directly
One of the main design choices used in this library is that it must be
possible to easily switch how input for the lexer-generator is build and
how (and what) is outputted by the code-generator. For this reason the
library allows you to first create a description of the lexer. This is
no more than a representation of an input file or a manually constructed
object structure. This description can then be used to create a
definition, which is also no more than an object structure. An algorithm
can use this definition directly (see tests for an example of this) or
code can be generated from this definition.

The programmer could use the (previously called) LexerDefinitionBuilder
to create a LexerDefinition. The programmer would create a
LexerDescriptionNode that would be used *internally* to create the
definition. This violated the idea of having these steps explicit.